backend/requirements.txt
fastapi
uvicorn
websockets
google-generativeai
python-dotenv